Move project status items to a new ui folder
