+++ extensions/typescript-language-features/src/typeScriptServiceClientHost.ts
@@ -17,11 +17,11 @@ import * as Proto from './protocol';
 import * as PConst from './protocol.const';
 import { OngoingRequestCancellerFactory } from './tsServer/cancellation';
 import { ILogDirectoryProvider } from './tsServer/logDirectoryProvider';
-import { ProjectStatus } from './tsServer/projectStatus';
 import { TsServerProcessFactory } from './tsServer/server';
 import { ITypeScriptVersionProvider } from './tsServer/versionProvider';
-import { VersionStatus } from './tsServer/versionStatus';
 import TypeScriptServiceClient from './typescriptServiceClient';
+import { ProjectStatus } from './ui/projectStatus';
+import { VersionStatus } from './ui/versionStatus';
 import { ActiveJsTsEditorTracker } from './utils/activeJsTsEditorTracker';
 import { coalesce, flatten } from './utils/arrays';
 import { ServiceConfigurationProvider } from './utils/configuration';
